requirements.txt
marimo
requests
pandas
anthropic
supabase
python-dotenv
aiohttp
nest_asyncio
resend
praw
selenium
webdriver-manager
